v2alpha1/pkg/api: sketch up a ttrpc plugin API proto.
Sketch up a prototype that would allow more versatile
NRI plugins to extend the functionality provided by a
runtime.

The prototype replaces the exec-based one-shot plugin
API with daemon-like plugins, where a single instance
of a plugin is responsible for handling all requests
intended for it. The JSON-based request/response API
on stdin/stdout is replaced with a ttrpc protocol.

The sketched up protocol allows plugins to hook into
the full CRI container lifecycle, to track the state
of both pods and containers, and to make (restricted)
changes to containers. These changes can happen both
as a response to certain lifecycle state transitions
or otherwise in an unsolicited fashion.

The sketched up protocol prototype allows plugins to
request changes to labels, annotations, devices, env
vars, mounts, assigned resources, and OCI hooks.