pkg/runtime: add runtime-interaction interface.
Add a package for interacting with plugins from
runtimes, along with a few additional supporting
packages.

This prototype gives up the original exec-based one-
shot plugin design. Instead it introduces daemon-like
plugins where a single plugin instance is responsible
for handling the full stream of requests and events
intended for that plugin.

The original JSON-based API using stdin/stdout as the
transport is replaced with a ttRPC-based protocol and
a unix-domain stream socket as a transport.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>