plugins: add demo cgroup hook hack.
Add a demo cgroup hook hack showing how an OCI hook can
be injected into a container at the createRuntime stage
to override the pods cgroupv2 cpu.weight setting.

Usage:

cgroup-hook-hack -idx 10 -cpu-weight 777 -pod test-hook-pod

This will override the cpu weight for 'test-hook-pod' with
weight 777. The plugin is dumb PoC. It will inject the hook
to each container in the pod and overriding the setting as
many times as there are containers in the pod...

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>