adaptation: blockable plugin synchronization.
Add an interface for blocking, re-allowing and requestin plugin
synchronization. These should be used, on the runtime side, to
block synchronization of registering plugins during pod and
container creation until the pod or container becomes visible
to listing. This should be done on the runtime side. On the
adaptation side use the corresponding functions to request
plugins synchronization during registration.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>