api: add seccomp adjustment
This adds an adjustment for seccomp policies. The intent is that people can
wholesale replace policies, or parse them, make some changes, and then send
them back. Sending them *to* NRI via containerd requires some containerd
patches as well, those are here: https://github.com/tych0/containerd/commits/nri-seccomp/

Specifically, we are interested in making the listenerPath of the policy
dynamic based on a k8s pod spec, so we can't use the Localhost custom
policy (well, we can use most of it, except for listenerPath, which we have
an NRI plugin to change based on this code).

This patch is a lot of boilerplate, which is unfortunate. There is a much
smaller but similar patch:
https://github.com/tych0/nri/commit/a70547a543f762542212fde804b9cd3f70cfc272
but it involves directly serializing a runtime-spec string

Finally, note the comment in generate.go: the runtime-tools generate code
does not have complete coverage for seccomp stuff, so I opted to not use
any of it, vs. adding more stuff to runtime-tools. The fact that there are
human and computer names is also confusing, it seems like we should stick
to the computer names for this particular interface.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>