api: make field owner lookup purely read-only.
Split per container FieldOwners lookup for plugin field claiming
and clearing from that of plugin field owner lookup so that the
formers create missing owners but the latter does not. This makes
field owner lookup a purely read-only operation, avoiding a data
read/write race when we have both builtin and external validators
present and we run validation in parallel on multiple goroutines.

Also, do not export the new ownersFor and mustOwnersFor functions
as they are never used externally.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>